refactor(advanced_dock_area): profile tools moved to separate module
